Implement Teams CRUD.

+++ app/controllers/teams_controller.rb
@@ -1,4 +1,5 @@
 class TeamsController < ApplicationController
+  respond_to :json
   # GET /teams
   def index
     @teams = Team.all
@@ -26,13 +27,25 @@ class TeamsController < ApplicationController
 
   # PUT/PATCH /teams/:id
   def update
+    team = Team.find params[:id]
+
+    if team.update(team_params)
+      render nothing: true, status: :no_content
+    else
+      render nothing: true, status: :not_found
+    end
   end
 
   # DELETE    /teams/:id
   def destroy
     team = Team.find params[:id]
-    team.destroy
-    respond_with team
+    if team
+      team.destroy
+      render nothing: true, status: :no_content
+      # render nothing: true, status: :ok
+    else
+      render nothing: true, status: :no_found
+    end
   end
 
 private
